refactor: move load_source_image_base64 to infrastructure storage (#251)
- Move load_source_image_base64 from app.presentation.helpers to app.infrastructure.storage.s3
- Update imports in solution_worker.py, exam_grading.py, and practice.py
- Preserve function signature, behavior, and S3 error handling exactly
- Remove stale S3 imports from presentation/helpers.py

Closes #251
