+++ backend/app/presentation/practice.py
@@ -26,7 +26,8 @@ from app.presentation.deps import (
     get_grading_vlm_client,
     get_s3_storage,
 )
-from app.presentation.helpers import build_problem_image_url, load_source_image_base64, parse_object_id
+from app.infrastructure.storage.s3 import load_source_image_base64
+from app.presentation.helpers import build_problem_image_url, parse_object_id
 from app.presentation.exam_helpers import problem_document_to_model
 from app.presentation.errors import ApiError
